// src/app/core/shared/models/service.model.ts
export interface Servics{
    service_id: number | any;
    vendor_id: number | any;
    business_id: number | any;
    business_name: string | any;
    service_name: string | any;
    description: string | any;
    minprice: number | any;
    maxprice: number | any;
    category: string | any;
    path: string | any;
    registration_date: string | any;
}